rename(FolderTitle): Change name Header to FolderTitle
